# store: send data immediately after getting it