Build a min-heap to immediately send back data to the querier after
getting it. Min heap is needed to maintain the invariant of Series()
that the returned data is sorted by the labels. My anecdotal data shows
that this could shave off a few percent of the total duration of queries
because we aren't doing one sorting pass at the end _after_ getting
_all_ of the data and we can free memory quicker.

I don't think any of the current benchmarks can show this difference?
It's probably the most visible when a Series() comes in which matches a
lot of series.

I have only done minimal testing - that I am still able to query data
via `/api/v1/query_range`.

Signed-off-by: Giedrius Statkevičius <giedrius.statkevicius@vinted.com>

* [ ] I added CHANGELOG entry for this change.
* [ ] Change is not relevant to the end user.

What do you think @bwplotka @kakkoyun @yeya24 ?