@bwplotka 2021-05-22
Thanks, overall there is a lot of room for improvement so great to see some movement here! (: 
 
So essentially we are replacing merge sort that joins data from blocks (series within block are sorted) with min heap?

I will review in June in more detail, but I would be against merging anything on this level without local Go benchmark that shows visible improvement. This is because it might surprising that we actually allocate here more than less etc 🙈