@GiedriusS 2021-05-24
As expected, CPU usage savings are great :scream: however there are some new memory allocations somewhere:
```
benchmark                                   old ns/op     new ns/op     delta
BenchmarkBlockSeries/concurrency:_1-16      9112258       3138514       -65.56%
BenchmarkBlockSeries/concurrency:_2-16      4902091       2117883       -56.80%
BenchmarkBlockSeries/concurrency:_4-16      2963023       1376239       -53.55%
BenchmarkBlockSeries/concurrency:_8-16      2247469       1054152       -53.10%
BenchmarkBlockSeries/concurrency:_16-16     2105249       861687        -59.07%
BenchmarkBlockSeries/concurrency:_32-16     3022499       832255        -72.46%

benchmark                                   old allocs     new allocs     delta
BenchmarkBlockSeries/concurrency:_1-16      31525          31546          +0.07%
BenchmarkBlockSeries/concurrency:_2-16      31352          21138          -32.58%
BenchmarkBlockSeries/concurrency:_4-16      32703          13817          -57.75%
BenchmarkBlockSeries/concurrency:_8-16      33396          10492          -68.58%
BenchmarkBlockSeries/concurrency:_16-16     32935          8976           -72.75%
BenchmarkBlockSeries/concurrency:_32-16     46036          9508           -79.35%

benchmark                                   old bytes     new bytes     delta
BenchmarkBlockSeries/concurrency:_1-16      5140092       16535258      +221.69%
BenchmarkBlockSeries/concurrency:_2-16      5295731       11614433      +119.32%
BenchmarkBlockSeries/concurrency:_4-16      5235786       8172685       +56.09%
BenchmarkBlockSeries/concurrency:_8-16      5442007       6578585       +20.89%
BenchmarkBlockSeries/concurrency:_16-16     6131926       5803753       -5.35%
BenchmarkBlockSeries/concurrency:_32-16     8115555       6078792       -25.10%
```

Trying to dig more into it