@GiedriusS 2021-06-17
Actually, this probably isn't worth it due to how Go's memory management works - even though we'd "free" the memory quicker, it doesn't reflect on the memory management side before the next GC pass. It would be best to have "full streaming" without a separate buffering & sorting step - then the overall gains from such improvement would be clearer. Thus, closing this (: